Automobile emissions data

This is a replicated 3^2 experiment reported in Box, Hunter, and Hunter (2005), Table 10.17.

Usage

codata

Format

A data frame with 18 observations on the following 3 variables.

x1

a numeric vector, coded design variable for ethanol concentration

x2

a numeric vector, coded design variable for air-to-fuel ratio

y

a numeric vector, the response (CO concentration, in micrograms per cubic meter)

Details

This example, when fitted with a second-order response surface, is an example of a rising ridge. The dataset is duscussed again one chapter later in the source text; Figure 11.17 of BH^2 suggests the coding formulas used in the example below.

Source

Box, GEP, Hunter, JS, and Hunter, WG (2005) Statistics for Experimenters (2nd ed), Wiley.

Examples

# Create a coded dataset based on info in BH^2 Fig 11.17
CO <- as.coded.data(codata,  x1 ~ (Ethanol - 0.2)/0.1,  x2 ~ A.F.ratio - 15)
names(CO)[3] <- "CO.conc"