Response-surface analysis

The rsm package provides functions useful for designing and analyzing experiments that are done sequentially in hopes of optimizing a response surface.

The function ccd can generate (and randomize) a central-composite design; it allows the user to specify an aliasing or fractional blocking structure. The function bbd generates and randomizes a Box-Behnken design. The function ccd.pick is useful for identifying good parameter choices in central-composite designs. Functions cube, star, foldover, dupe, and djoin are also provided to build-up designs from individual blocks. The function varfcn allows the experimenter to examine the predictive capabilities of a design before collecting data.

The function rsm is an enhancement of lm that provides for additional analyses peculiar to response surfaces. It requires a model formula that contains a call to FO or SO to specify a first- or second-order model. Once the model is fitted, the steepest function may be used to obtain the direction of steepest ascent (or descent). canonical.path is an alternative to steepest for second-order response surfaces.

In RSM methods, appropriate coding of data is important not only for numerical stability, but for proper scaling of results; the function coded.data and its relatives facilitate this coding requirement.

Finally, a few more functions are provided that may be useful beyond response-surface applications. contour.lm, persp.lm, and image.lm aids in visualizing a response surface, or of any other lm object where a surface is fitted. model.data recovers the data used in a lm call, but unlike model.frame, no polynomials, factors, etc. are expanded.

For more information and examples, use vignette("rsm") and vignette("rs-illus"). Additionally, vignette("rsm-plots") provides some illustrations of the graphics functions.

Author

Russell V. Lenth

Maintainer: Russell V. Lenth <russell-lenth@uiowa.edu>

References

Box, GEP, Hunter, JS, and Hunter, WG (2005) Statistics for Experimenters (2nd ed.), Wiley-Interscience.

Lenth RV (2009) “Response-Surface Methods in R, Using rsm”, Journal of Statistical Software, 32(7), 1–17. doi:10.18637/jss.v032.i07

Myers, RH, Montgomery, DC, and Anderson-Cook, CM (2009), Response Surface Methodology (3rd ed.), Wiley.