Chemical Reaction Data

These data are from a central composite design with 2 factors in 2 blocks. The design variables are in actual, not coded, form.

Usage

ChemReact
ChemReact1
ChemReact2

Format

A data frame with 14 observations on the following 4 variables.

Time

a numeric vector; design variable with settings of 80, 85, and 90.

Temp

a numeric vector; design variable with settings of 170, 175, and 180.

Block

a factor with levels B1 B2. Block B1 is a first-order design with 3 center points. Block B2 consists of axis points and 3 more center points.

Yield

a numeric vector; response variable: yield of the chemical process.

ChemReact1 and ChemReact2 are the separate blocks. Each has 7 runs and three variables (Block is excluded from these).

Source

Table 7.6 of Myers, RH, Montgomery, DC, and Anderson-Cook, CM (2009), Response Surface Methodology (3rd ed.), Wiley.